- add internet Package page

+++ app/src/main/java/com/technotapp/servicestation/adapter/DataModel/WalletAdapterModel.java
@@ -3,7 +3,7 @@ package com.technotapp.servicestation.adapter.DataModel;
 public class WalletAdapterModel {
     public String title;
     public boolean isActivePin;
-    public int accountNumber;
+    public long accountNumber;
     public int nidAccountType;
 
 }
